HBASE-6865: Snapshot File Cleaner (Jesse Yates)
Adding snapshot file cleaners (or rather file cleaner delegates) to keep hfiles and logs files in a snapshot from being removed.

This is significantly different than the posted snapshot-cleaners branch on github as it favors moving to a more composition (rather than inheritance-oriented) based layout. The main class is the SnasphotFileCache which tries to intelligently cache file names. The cleaners just supply a directory filter (so hfile cleaner filter is on the regions, log cleaner is on the .logs directory), and the cache keeps itself updated and periodically refreshes to remove deleted snapshots.